requirements.txt
django==4.2.4
djangorestframework==3.14.0
django-environ==0.10.0
drf-spectacular==0.26.4
psycopg2-binary==2.9.7